Restructure code to prevent long files
